리사이클러뷰 확인

+++ app/src/main/java/com/example/picknumber/model/SearchModel/Search.kt
@@ -3,6 +3,6 @@ package com.example.picknumber.model.SearchModel
 import java.io.Serializable
 
 data class Search(
-    var distance: Int,
+    var distance: Double,
     val bankName: String,
 )
